+++ src/js/ajax_utils.js
@@ -1,10 +1,10 @@
-function ajaxCheckPid(base, pid, interval) {
+function ajaxCheckPid(base, method, pid, interval) {
     var limit = 30000; // in miliseconds
     var wait  = parseInt(interval);
     if (wait*2 < limit) {
         wait  = wait*2;
     } else { wait = limit; }
-    new Ajax.Updater('response', base+'/check_pid', 
+    new Ajax.Updater('response', base+'/'+method,
     { method: 'get' ,
       parameters : {'pid': pid},
       onException: function() {return;},
@@ -15,7 +15,7 @@ function ajaxCheckPid(base, pid, interval) {
         // if check_pid still working on request, call again
         if (transport.responseText.match(/processing PID/)) {
             transport.responseText += msg;
-            setTimeout('ajaxCheckPid("'+base+'","'+pid+'","'+wait+'")', wait);
+            setTimeout('ajaxCheckPid("'+base+'","'+method+'","'+pid+'","'+wait+'")', wait);
         }
       }
     });